Refactor middlewares anime endpoint
