+++ middlewares/anime/controllers/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+  getAll: require('./getAll'),
+  getById: require('./getById'),
+  addAnime: require('./addAnime'),
+  deleteAnimeById: require('./deleteAnimeById'),
+  deleteAllAnime: require('./deleteAllAnime'),
+  updateAnimeData: require('./updateAnimeData'),
+}
